feat: short UUID prefix support for agu/respondi + split 794-line service.py monolith (fixes #38)
Feature:
- get_email() tries exact match, then prefix fallback via
  find_message_by_uuid_prefix() or LIKE query on direct DB fallback
- Handles 0/multiple matches gracefully with warning messages

Refactor:
- Split 794-line service.py monolith into service/ package (6 files,
  all <250 lines) using A-lien's mixin pattern:
  - __init__.py: public API, singleton
  - helpers.py: EmailSummary, ActionSuggestion, _extract_json
  - registry.py: cross-module service lazy-loaders
  - agent_service.py: AgentService base (init, email fetching)
  - summarization_mixin.py: summarize_emails, generate_reply, extract_actions
  - action_mixin.py: create_calendar_event, create_todo, create_knowledge_entry
